@@ -2,15 +2,12 @@
 import { subclass, property } from 'esri/core/accessorSupport/decorators';
 import Widget from 'esri/widgets/Widget';
 import DetailPanel from './DetailPanel';
-import i18n = require('dojo/i18n!../nls/resources');
-import { tsx, renderable } from 'esri/widgets/support/widget';
-import { setPageTitle } from 'ApplicationBase/support/domHelper';
+import { tsx } from 'esri/widgets/support/widget';
+import { setPageTitle } from 'TemplatesCommonLib/baseClasses/support/domHelper';
 import { init } from "esri/core/watchUtils";
 
-import ConfigurationSettings = require('../ConfigurationSettings');
+import ConfigurationSettings from '../ConfigurationSettings';
 import esri = __esri;
-
-
 
 const CSS = {
 	theme: 'app-header',
@@ -28,6 +25,7 @@ const CSS = {
 
 interface HeaderProps extends esri.WidgetProperties {
 	config: ConfigurationSettings,
+	sharedTheme: any,
 	detailPanel: DetailPanel
 }
 
@@ -40,9 +38,9 @@ class Header extends (Widget) {
 	//--------------------------------------------------------------------------
 
 	@property()
-	@renderable(["title", "titleLink"])
 	config: ConfigurationSettings;
 	@property() detailPanel: DetailPanel;
+	@property() sharedTheme: any;
 	//--------------------------------------------------------------------------
 	//
 	//  Public Methods
@@ -50,52 +48,88 @@ class Header extends (Widget) {
 	//--------------------------------------------------------------------------
 	constructor(props: HeaderProps) {
 		super(props);
+		this._onTitleUpdate = this._onTitleUpdate.bind(this);
 	}
 	postInitialize() {
-		const handle = init(this, "config.title", this._onTitleUpdate);
+		this.own(init(this, "config.title", this._onTitleUpdate));
 
-		this.own(handle);
 	}
 	render() {
-		const { title, titleLink } = this.config;
-
+		const { title, titleLink, headerBackground, headerColor, theme, enableHeaderBackground, enableHeaderColor, showIntroduction } = this.config;
+		const customStyle = this._createCustomStyle();
 		const titleNode = titleLink ? (
-			<a target="_blank" rel="noopener" href={titleLink}>
+			<a target="_blank" rel="noopener noreferer" href={titleLink}>
 				{title}
 			</a>
 		) : (
-				title
-			);
-		const infoButton = this.detailPanel ? <div class="right">
-			<button bind={this} onclick={this._showDetailPanel} aria-label={i18n.tools.info} title={i18n.tools.info} class={this.classes(CSS.theme, CSS.calciteStyles.topNavTitle, CSS.calciteStyles.buttonLink, CSS.calciteStyles.iconDesc)}></button>
-		</div> : null;
-
-
+			title
+		);
+		const headerClasses = (headerBackground && enableHeaderBackground) || (headerColor && enableHeaderColor) ? [CSS.calciteStyles.topNavTitle] : [CSS.calciteStyles.ellipsis, CSS.calciteStyles.topNavTitle]
+		const color = theme === "light" ? "inverse" : "neutral";
+		const infoButton = showIntroduction && this.detailPanel ? <div class="info-button" key="info-button">
+			<calcite-button
+				appearance="transparent"
+				color={color}
+				bind={this}
+				id="infoButton"
+				onclick={this._showDetailPanel}
+				aria-label={this.config.appBundle.tools.info}
+				icon-start="information-f"
+			>
+			</calcite-button>
+		</div > : null;
 		return (
-			<header class={this.classes(CSS.calciteStyles.topNav, CSS.theme)}>
-				<div class={this.classes(CSS.calciteStyles.fade)}>
-					<h1 title={title} class={this.classes(CSS.calciteStyles.topNavTitle, CSS.calciteStyles.ellipsis)}>{titleNode}</h1>
-				</div>
+			<header theme={theme} style={customStyle} class={this.classes(CSS.calciteStyles.topNav, CSS.theme)}>
+				<h1 title={title} class={this.classes(headerClasses)}>{titleNode}</h1>
 				{infoButton}
-
 			</header>
 		);
+	}
+	_createCustomStyle() {
+		const hasSharedTheme = document.body.classList.contains("shared-theme");
+
+		const { headerBackground, applySharedTheme, enableHeaderBackground, enableHeaderColor, headerColor, theme } = this.config;
+
+
+		// Default theme colors 
+		let backgroundColor = theme === "light" ? "#4a4a4a" : "#323232";
+		let textColor = "#fff";
+
+		// Overwrite defaults above with shared themes 
+		if (hasSharedTheme && applySharedTheme) {
+			const { header } = this.sharedTheme;
+			backgroundColor = header?.background ? header.background : backgroundColor;
+			textColor = header?.text ? header.text : textColor;
+		}
+
+		// These always win (overwrite theme and shared theming)
+		backgroundColor = headerBackground && enableHeaderBackground ? headerBackground : backgroundColor;
+
+		textColor = headerColor && enableHeaderColor ? headerColor : textColor;
+		const calciteVariable = theme === "dark" ? `--calcite-ui-text-1` : `--calcite-ui-foreground-1`;
+		return `
+		${calciteVariable}:${textColor};
+		background-color:${backgroundColor};
+		fill:${textColor};
+		color:${textColor};`;
+
 	}
 	_showDetailPanel(e) {
 		// add class to detail panel to add close button and position at top of panel
 		const node = this.detailPanel.container as HTMLElement;
-		node.classList.add("info-triggered");
 
-		// Show panel unless we click button with panel already open
 		if (node && !node.classList.contains("hide")) {
 			this.detailPanel.hidePanel();
 		} else {
 			this.detailPanel.showPanel();
 		}
+
 	}
 
 	private _onTitleUpdate() {
-		setPageTitle(this.config.title);
+		if (this?.config?.title) {
+			setPageTitle(this.config.title);
+		}
 	};
 }
 export = Header;
