@@ -2,21 +2,18 @@
 import { subclass, property } from "esri/core/accessorSupport/decorators";
 import Widget from "esri/widgets/Widget";
 import Handles from "esri/core/Handles";
-import i18n = require("dojo/i18n!../nls/resources");
 import { tsx } from "esri/widgets/support/widget";
 import { getLocale } from "esri/intl";
 import Slider from "esri/widgets/Slider";
 import esri = __esri;
 
-import ConfigurationSettings = require("../ConfigurationSettings");
+import ConfigurationSettings from "../ConfigurationSettings";
 
 
 const CSS = {
     panel: "panel",
     panelNoBorder: "panel-no-border",
     refinePanel: "panel-refine-results",
-    filterButton: "filter-button",
-    filterIcon: "icon-ui-filter",
     buttonLink: "btn-link",
     button: "btn",
     hide: "hide"
@@ -44,7 +41,7 @@ class RefineResults extends (Widget) {
     //--------------------------------------------------------------------------
 
     _handles: Handles = new Handles();
-    _filterPanelNode: HTMLElement = null;
+
     //--------------------------------------------------------------------------
     //
     //  Public Methods
@@ -54,22 +51,27 @@ class RefineResults extends (Widget) {
         super(props);
     }
     render() {
+        const { enableBufferSearch, singleLocationPolygons } = this.config;
+        const slider = enableBufferSearch && !singleLocationPolygons ? <div bind={this} afterCreate={this._createSlider} /> : null;
         return (
-            <div>
-                <div class={this.classes(CSS.panel, CSS.refinePanel, CSS.panelNoBorder)}>
-                    <div bind={this} afterCreate={this._createSlider} />
-                </div>
+            <div class={this.classes(CSS.panel, CSS.refinePanel, CSS.panelNoBorder)}>
+                {slider}
             </div>
         );
     }
-    _createSlider(node: HTMLElement) {
-        const { searchUnits, sliderRange, precision, inputsEnabled } = this.config;
-        const { minimum, maximum } = sliderRange;
-        const distanceSlider = new Slider({
+
+    _createSlider(container: HTMLElement) {
+        const { searchUnits, sliderRange, precision, inputsEnabled, showSlider } = this.config;
+
+        let sliderDefault = sliderRange?.default || 1.2;
+        let minimum = sliderRange.minimum || 0;
+        let maximum = sliderRange.maximum || 2;
+
+        if (!showSlider) return;
+        this.slider = new Slider({
             min: minimum,
             max: maximum,
-            values: [sliderRange.default],
-            precision: parseInt(precision),
+            values: [sliderDefault],
             visibleElements: {
                 labels: true,
                 rangeLabels: true
@@ -77,14 +79,21 @@ class RefineResults extends (Widget) {
             labelInputsEnabled: inputsEnabled,
             rangeLabelInputsEnabled: inputsEnabled,
             snapOnClickEnabled: true,
-            container: node
+            inputCreatedFunction: (input, type, index) => {
+                input.setAttribute("type", "number");
+                input.setAttribute("pattern", "[0-9]*");
+            },
+            precision,
+            container
         });
-        this.slider = distanceSlider;
+        let locale = getLocale();
+        if (!locale) locale = "en";
 
-        const locale = getLocale();
         // Append units to range labels   
+        if (!this.slider) return;
         this.slider.labelFormatFunction = (value: number, type: string) => {
-            let convertedUnits = i18n.units[searchUnits];
+            let convertedUnits = this.config.appBundle.units[searchUnits];
+
             convertedUnits = convertedUnits ? convertedUnits.abbr : "mi";
             //type can be min, max, value 
             if (type == "value") {
@@ -94,33 +103,29 @@ class RefineResults extends (Widget) {
             }
         };
         // Take locale into account
-        distanceSlider.inputFormatFunction = (value) => {
+        this.slider.inputFormatFunction = (value) => {
             return value.toLocaleString(locale);
         };
 
         this._handles.remove(["slider,adjust"]);
-        this._handles.add(distanceSlider.watch("values", (value) => {
-            // this.config.distance = value;
-            if (distanceSlider.state === "ready") {
+        this._handles.add(this.slider.watch("values", (value) => {
+            if (this.slider.state === "ready") {
                 this.value = value[0];
             }
         }), "slider");
-        distanceSlider.on("thumb-drag", (e) => {
+        this.slider.on("thumb-drag", (e) => {
             if (e.state === "stop") {
                 this.value = e.value;
             }
         });
         this._adjustLabels();
-        this._handles.add(distanceSlider.watch(["max", "min"], () => this._adjustLabels), "adjust");
-
-        return distanceSlider;
+        this._handles.add(this.slider.watch(["max", "min"], () => this._adjustLabels), "adjust");
+        return this.slider;
     }
     destroy() {
         this._handles.removeAll();
     }
-    _storeNode(node: HTMLElement) {
-        this._filterPanelNode = node;
-    }
+
 
     _adjustLabels() {
         // Move label down so it doesn't look so smushed 
@@ -140,22 +145,22 @@ class RefineResults extends (Widget) {
                 maxInput?.classList.remove("bottom-label");
             }
         }
-
     }
     public updateSliderProps(propertyName: string, value: any) {
 
         if (this.slider) {
-
             if (propertyName === "sliderRange") {
                 if (value?.default !== this.slider.values[0]) {
                     this.slider.values = [this.config.sliderRange.default];
                 }
             }
-
+            if (propertyName === "precision") {
+                this.slider.precision = value ?? 4;
+            }
             if (propertyName === "sliderRange" && value?.maximum !== this.slider.max) {
                 this.slider.max = this.config.sliderRange?.maximum;
             }
-            if (propertyName === "minDistance" && value?.minimum !== this.slider.min) {
+            if (propertyName === "sliderRange" && value?.minimum !== this.slider.min) {
                 this.slider.min = this.config.sliderRange?.minimum;
             }
             if (propertyName === "inputsEnabled") {
@@ -169,7 +174,7 @@ class RefineResults extends (Widget) {
                 if (this.slider) {
                     // Append units to range labels   
                     this.slider.labelFormatFunction = (value: number, type: string) => {
-                        let convertedUnits = i18n.units[searchUnits];
+                        let convertedUnits = this.config.appBundle.units[searchUnits];
                         convertedUnits = convertedUnits ? convertedUnits.abbr : "mi";
 
                         if (type === "value") {
