@@ -10,29 +10,10 @@
   See the License for the specific language governing permissions and
   limitations under the License.​
 */
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = function (d, b) {
-        extendStatics = Object.setPrototypeOf ||
-            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-        return extendStatics(d, b);
-    };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
-var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
-    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
-    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
-    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
-    return c > 3 && r && Object.defineProperty(target, key, r), r;
-};
-define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/core/Accessor"], function (require, exports, decorators_1, Accessor) {
+define(["require", "exports", "tslib", "esri/core/accessorSupport/decorators", "esri/core/Accessor"], function (require, exports, tslib_1, decorators_1, Accessor) {
     "use strict";
     var ShareFeatures = /** @class */ (function (_super) {
-        __extends(ShareFeatures, _super);
+        tslib_1.__extends(ShareFeatures, _super);
         function ShareFeatures() {
             var _this = _super !== null && _super.apply(this, arguments) || this;
             _this.embedMap = true;
@@ -61,23 +42,22 @@ define(["require", "exports", "esri/core/accessorSupport/decorators", "esri/core
             enumerable: false,
             configurable: true
         });
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property({ value: true })
         ], ShareFeatures.prototype, "copyToClipboard", null);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property({ value: true })
         ], ShareFeatures.prototype, "shareServices", null);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property()
         ], ShareFeatures.prototype, "embedMap", void 0);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property()
         ], ShareFeatures.prototype, "shortenLink", void 0);
-        ShareFeatures = __decorate([
+        ShareFeatures = tslib_1.__decorate([
             decorators_1.subclass('ShareFeatures')
         ], ShareFeatures);
         return ShareFeatures;
     }((Accessor)));
     return ShareFeatures;
 });
-//# sourceMappingURL=ShareFeatures.js.map
