@@ -1,59 +1,4 @@
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = function (d, b) {
-        extendStatics = Object.setPrototypeOf ||
-            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-        return extendStatics(d, b);
-    };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
-var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
-    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
-    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
-    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
-    return c > 3 && r && Object.defineProperty(target, key, r), r;
-};
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-define(["require", "exports", "esri/core/Accessor", "esri/core/Collection", "esri/core/accessorSupport/decorators", "esri/geometry/Point", "esri/request", "./ShareItem", "./ShareFeatures", "esri/geometry/projection", "esri/geometry/SpatialReference"], function (require, exports, Accessor, Collection, decorators_1, Point, esriRequest, ShareItem, ShareFeatures, projection, SpatialReference) {
+define(["require", "exports", "tslib", "esri/core/Accessor", "esri/core/Collection", "esri/core/accessorSupport/decorators", "esri/geometry/Point", "esri/request", "./ShareItem", "./ShareFeatures", "esri/geometry/projection", "esri/geometry/SpatialReference"], function (require, exports, tslib_1, Accessor, Collection, decorators_1, Point, esriRequest, ShareItem, ShareFeatures, projection, SpatialReference) {
     "use strict";
     //----------------------------------
     //
@@ -93,7 +38,7 @@ define(["require", "exports", "esri/core/Accessor", "esri/core/Collection", "esr
     //----------------------------------
     var SHORTEN_API = "https://arcg.is/prod/shorten";
     var ShareViewModel = /** @class */ (function (_super) {
-        __extends(ShareViewModel, _super);
+        tslib_1.__extends(ShareViewModel, _super);
         function ShareViewModel() {
             //----------------------------------
             //
@@ -194,9 +139,9 @@ define(["require", "exports", "esri/core/Accessor", "esri/core/Collection", "esr
         //
         //----------------------------------
         ShareViewModel.prototype.generateUrl = function () {
-            return __awaiter(this, void 0, void 0, function () {
+            return tslib_1.__awaiter(this, void 0, void 0, function () {
                 var url, shortenLink, shortenedUrl;
-                return __generator(this, function (_a) {
+                return tslib_1.__generator(this, function (_a) {
                     switch (_a.label) {
                         case 0: return [4 /*yield*/, this._generateShareUrl()];
                         case 1:
@@ -221,9 +166,9 @@ define(["require", "exports", "esri/core/Accessor", "esri/core/Collection", "esr
         //
         //----------------------------------
         ShareViewModel.prototype._generateShareUrl = function () {
-            return __awaiter(this, void 0, void 0, function () {
+            return tslib_1.__awaiter(this, void 0, void 0, function () {
                 var href, _a, x, y, spatialReference, centerPoint, point;
-                return __generator(this, function (_b) {
+                return tslib_1.__generator(this, function (_b) {
                     switch (_b.label) {
                         case 0:
                             href = window.location.href;
@@ -247,9 +192,9 @@ define(["require", "exports", "esri/core/Accessor", "esri/core/Collection", "esr
             });
         };
         ShareViewModel.prototype._processPoint = function (point) {
-            return __awaiter(this, void 0, void 0, function () {
+            return tslib_1.__awaiter(this, void 0, void 0, function () {
                 var _a, isWGS84, isWebMercator, point_1, outputSpatialReference, projectedPoint;
-                return __generator(this, function (_b) {
+                return tslib_1.__generator(this, function (_b) {
                     switch (_b.label) {
                         case 0:
                             _a = point.spatialReference, isWGS84 = _a.isWGS84, isWebMercator = _a.isWebMercator;
@@ -313,9 +258,9 @@ define(["require", "exports", "esri/core/Accessor", "esri/core/Collection", "esr
             return shareParams;
         };
         ShareViewModel.prototype._shorten = function (url) {
-            return __awaiter(this, void 0, void 0, function () {
+            return tslib_1.__awaiter(this, void 0, void 0, function () {
                 var request, shortUrl;
-                return __generator(this, function (_a) {
+                return tslib_1.__generator(this, function (_a) {
                     switch (_a.label) {
                         case 0:
                             this._shortening = true;
@@ -342,42 +287,41 @@ define(["require", "exports", "esri/core/Accessor", "esri/core/Collection", "esr
         ShareViewModel.prototype._roundValue = function (val) {
             return parseFloat(val.toFixed(4));
         };
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property({
                 dependsOn: ["view.ready"],
                 readOnly: true
             })
         ], ShareViewModel.prototype, "state", null);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property({
                 dependsOn: ["shareUrl"],
                 readOnly: true
             })
         ], ShareViewModel.prototype, "embedCode", null);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property()
         ], ShareViewModel.prototype, "view", void 0);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property()
         ], ShareViewModel.prototype, "shareModalOpened", void 0);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property({
                 type: ShareItemCollection
             })
         ], ShareViewModel.prototype, "shareItems", void 0);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property({
                 type: ShareFeatures
             })
         ], ShareViewModel.prototype, "shareFeatures", void 0);
-        __decorate([
+        tslib_1.__decorate([
             decorators_1.property({ readOnly: true })
         ], ShareViewModel.prototype, "shareUrl", void 0);
-        ShareViewModel = __decorate([
+        ShareViewModel = tslib_1.__decorate([
             decorators_1.subclass("ShareViewModel")
         ], ShareViewModel);
         return ShareViewModel;
     }((Accessor)));
     return ShareViewModel;
 });
-//# sourceMappingURL=ShareViewModel.js.map
